Add z_top_clamp part
This allows a threaded rod to be mounted across the top of the Z rods.
This stabilizes the Z axis in the X direction, and can also be used as
a handle, making it much easier to transport the machine. The clamp has
a gusset joint for added strength, and a new parameter has been added
to control the size of the gusset.