Huge refactoring, add platform and rod measurements
This update is not fully tested. It is a huge refactoring of the code
to consolidate the formulae that were being used repeatedly and try to
make it easier to read. It also produces a DFX that can be used to
laser cut a platform and calculates rod lengths based on the desired
build volume and provide a more complete visualization.

This should not change the shapes of any of the existing parts. If it
does, it's a bug.